requirements.txt
numpy==2.2.6
fastapi==0.115.12
librosa==0.11.0
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
Pillow==11.2.1
pydantic==2.11.5
pydantic_settings==2.9.1
PyYAML==6.0.1
Requests==2.32.3
scipy==1.15.3
scikit-image==0.25.2
soundfile==0.13.1
torch >=1.3
torchvision
tqdm==4.64.1
wandb==0.20.0
websockets==15.0.1
pybase64==1.4.1
uvicorn==0.34.3
jsons==1.6.3
wave==0.0.2
av==14.4.0